hw06-solid completed
Refactoring according to the comments and remarks of the teacher

+++ hw06-solid/src/main/java/ru/otus/atm/interaction/Atm.java
@@ -4,11 +4,10 @@ import ru.otus.atm.interaction.exception.AtmInteractionException;
 import ru.otus.atm.money.Banknote;
 import ru.otus.atm.money.Currency;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface Atm {
     List<Banknote> requestMoney(Currency currency, String amount) throws AtmInteractionException;
     void insertMoney(List<Banknote> banknotes);
-    BigInteger getBalanceForCurrency(Currency currency);
+    Integer getBalanceForCurrency(Currency currency);
 }
