+++ hw06-solid/src/main/java/ru/otus/atm/mechanism/cell/Cell.java
@@ -3,12 +3,10 @@ package ru.otus.atm.mechanism.cell;
 import ru.otus.atm.money.Banknote;
 import ru.otus.atm.mechanism.cell.exception.CellOperationException;
 
-import java.math.BigInteger;
-
 public interface Cell<T extends Banknote> {
     T extract() throws CellOperationException;
     void insertBanknote(T banknote) throws CellOperationException;
     int getBanknotesAmount();
 
-    BigInteger getDenomination();
+    Integer getDenomination();
 }
