+++ hw06-solid/src/main/java/ru/otus/atm/mechanism/cell/factory/CellCreator.java
@@ -4,9 +4,7 @@ import ru.otus.atm.money.Banknote;
 import ru.otus.atm.mechanism.cell.Cell;
 import ru.otus.atm.money.Currency;
 
-import java.math.BigInteger;
-
 public abstract class CellCreator<T extends Banknote> {
     protected static final int DEFAULT_CELL_SIZE = 2500;
-    abstract public Cell<T> create(BigInteger denomination, Currency currency);
+    abstract public Cell<T> create(Integer denomination, Currency currency);
 }
