+++ hw06-solid/src/main/java/ru/otus/atm/mechanism/cluster/ClusterManager.java
@@ -4,13 +4,12 @@ import ru.otus.atm.money.Banknote;
 import ru.otus.atm.money.Currency;
 import ru.otus.atm.mechanism.cluster.exception.ClusterOperationException;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface ClusterManager<T extends Banknote> {
     void putBanknote(List<Banknote> banknotes) throws ClusterOperationException;
-    List<T> requestBanknote(BigInteger nominal, int amount) throws ClusterOperationException;
-    int getBanknoteAmountFromCell(BigInteger denomination);
-    BigInteger getTotalSumFromCluster();
+    List<T> requestBanknote(Integer nominal, int amount) throws ClusterOperationException;
+    int getBanknoteAmountFromCell(Integer denomination);
+    Integer getTotalSumFromCluster();
     Currency getClusterCurrency();
 }
